Security improvements
Addressing the #2 security concerns reported by @tvibliani about the user created with the script. This is made that way because this is a “system” user. It is there to own and run the application, it isn’t supposed to be a person type user with a login etc. In Ubuntu, a system user gets a UID below 1000, has no shell (it’s actually /bin/false) and has logins disabled.