Deleting sudoer acces
The odoo user is a system user, created even without shell access rights and no password, because his solely purpose is to run the whole odoo process and that's all. The odoo user should not have any access level to the system beyond the scope of his own software.

Sudoers access should only be given to known system administrators. It adds an small overhead, but it's worth it as an small exchange from a good security policy point of view.